+++ src/main/java/me/jellysquid/mods/sodium/mixin/core/MixinFrustum.java
@@ -3,6 +3,7 @@ package me.jellysquid.mods.sodium.mixin.core;
 import me.jellysquid.mods.sodium.client.util.frustum.FrustumAccessor;
 import net.minecraft.client.render.Frustum;
 import org.joml.FrustumIntersection;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -24,8 +25,8 @@ public abstract class MixinFrustum implements FrustumAccessor {
     private double z;
 
     @Override
-    public Vector3f getTranslation() {
-        return new Vector3f((float) this.x, (float) this.y, (float) this.z);
+    public Vector3d getTranslation() {
+        return new Vector3d(this.x, this.y, this.z);
     }
 
     @Override
